Cannot read property 'onWillCloseEditor' of undefined (fix #99682)
